requirements.txt
tox
pre-commit
django-birthday
jeraconv
jaconv
